tried styling, add resolves to webpack, seperated graphql files
