+++ client/components/Header.jsx
@@ -4,14 +4,31 @@ import PropTypes from 'prop-types';
 const Header = (props) => {
   const { username, handleLogOut } = props;
 
+  const headerStyles = {
+    height: '20px',
+    backgroundColor: '#0000c4',
+    padding: '20px',
+    color: '#FFF',
+    display: 'flex',
+  };
+
+  const titleStyles = {
+    flexGrow: 11,
+  };
+
+  const menuStyles = {
+    flexGrow: 1,
+  };
+
   return (
-    <div id="header">
+    <div id="header" style={headerStyles}>
       <div
         id="title"
+        style={titleStyles}
       >
         Cool Chat
       </div>
-      <div>
+      <div id="menu" style={menuStyles}>
         <div className="username">
           { username }
         </div>
