+++ client/components/Sidebar.jsx
@@ -1,9 +1,16 @@
 import React, { Component } from 'react';
 import ActiveUser from './ActiveUser.jsx';
 
+const sidebarStyles = {
+  flexGrow: 1,
+  backgroundColor: 'Black',
+  height: '100%',
+  // height: 'calc(100vh - 40px)',
+};
+
 function Sidebar({ activeUsers }) {
   return (
-    <div className="sidebar">
+    <div className="sidebar" style={sidebarStyles}>
       <h3>Active Users</h3>
       <div>
         {activeUsers.reduce((usersList, curUser) => {
