+++ client/index.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { render } from 'react-dom';
+import { ApolloClient } from 'apollo-boost';
+import { ApolloProvider } from 'react-apollo';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { HttpLink } from 'apollo-link-http';
+import App from './App';
+
+
+const cache = new InMemoryCache();
+const link = new HttpLink({
+  uri: 'http://localhost:4000/graphql',
+});
+
+const client = new ApolloClient({
+  cache,
+  link,
+});
+
+render(
+  <ApolloProvider client={client}>
+    <App />
+  </ApolloProvider>,
+  document.querySelector('#root'),
+);
