+++ client/schema/queries.js
@@ -1,0 +1,17 @@
+import gql from 'graphql-tag';
+
+export const messageQuery = gql`
+  {
+    messages {
+      username
+      message
+      created_at
+    }
+  }
+`;
+
+export const userQuery = gql`
+  {
+    users
+  }
+`;
