+++ index.html
@@ -1,14 +1,27 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=`, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500">
+  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
   <title>Cool Chat for Cool Kids</title>
 
 </head>
+
 <body>
   <div id="root"></div>
   <script src="dist/bundle.js"></script>
 </body>
+
 </html>
+
+<style>
+  html, body, #root {
+    margin: 0;
+    padding: 0;
+    height: 100%;
+  }
+</style>
