+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
 
-  entry: './client/index.js',
+  entry: './client/index.jsx',
   mode: 'development',
   output: {
     path: path.resolve(__dirname, 'dist'),
@@ -26,6 +26,9 @@ module.exports = {
         },
       },
     ],
+  },
+  resolve: {
+    extensions: ['.js', '.jsx'],
   },
   devServer: {
     publicPath: '/dist/',
